refactor: 优化代码组织结构
