+++ service/singleton/singleton.go
@@ -1,18 +1,13 @@
 package singleton
 
 import (
-	"fmt"
-	"sort"
-	"sync"
+	"gorm.io/driver/sqlite"
 	"time"
 
 	"github.com/patrickmn/go-cache"
-	"github.com/robfig/cron/v3"
 	"gorm.io/gorm"
 
 	"github.com/naiba/nezha/model"
-	"github.com/naiba/nezha/pkg/utils"
-	pb "github.com/naiba/nezha/proto"
 )
 
 var Version = "v0.12.18" // ！！记得修改 README 中的 badge 版本！！
@@ -22,86 +17,52 @@ var (
 	Cache *cache.Cache
 	DB    *gorm.DB
 	Loc   *time.Location
-
-	ServerList map[uint64]*model.Server // [ServerID] -> model.Server
-	SecretToID map[string]uint64        // [ServerSecret] -> ServerID
-	ServerLock sync.RWMutex
-
-	SortedServerList []*model.Server // 用于存储服务器列表的 slice，按照服务器 ID 排序
-	SortedServerLock sync.RWMutex
 )
 
-// Init 初始化时区为上海时区
+// Init 初始化singleton
 func Init() {
+	// 初始化时区至上海 UTF+8
 	var err error
 	Loc, err = time.LoadLocation("Asia/Shanghai")
 	if err != nil {
 		panic(err)
 	}
+
+	Conf = &model.Config{}
+	Cache = cache.New(5*time.Minute, 10*time.Minute)
 }
 
-// ReSortServer 根据服务器ID 对服务器列表进行排序（ID越大越靠前）
-func ReSortServer() {
-	ServerLock.RLock()
-	defer ServerLock.RUnlock()
-	SortedServerLock.Lock()
-	defer SortedServerLock.Unlock()
+// LoadSingleton 加载子服务并执行
+func LoadSingleton() {
+	LoadNotifications() // 加载通知服务
+	LoadServers()       // 加载服务器列表
+	LoadCronTasks()     // 加载定时任务
+}
 
-	SortedServerList = []*model.Server{}
-	for _, s := range ServerList {
-		SortedServerList = append(SortedServerList, s)
+// InitConfigFromPath 从给出的文件路径中加载配置
+func InitConfigFromPath(path string) {
+	err := Conf.Read(path)
+	if err != nil {
+		panic(err)
 	}
+}
 
-	// 按照服务器 ID 排序的具体实现（ID越大越靠前）
-	sort.SliceStable(SortedServerList, func(i, j int) bool {
-		if SortedServerList[i].DisplayIndex == SortedServerList[j].DisplayIndex {
-			return SortedServerList[i].ID < SortedServerList[j].ID
-		}
-		return SortedServerList[i].DisplayIndex > SortedServerList[j].DisplayIndex
+// InitDBFromPath 从给出的文件路径中加载数据库
+func InitDBFromPath(path string) {
+	var err error
+	DB, err = gorm.Open(sqlite.Open(path), &gorm.Config{
+		CreateBatchSize: 200,
 	})
-}
-
-// =============== Cron Mixin ===============
-
-var CronLock sync.RWMutex
-var Crons map[uint64]*model.Cron
-var Cron *cron.Cron
-
-func ManualTrigger(c model.Cron) {
-	CronTrigger(c)()
-}
-
-func CronTrigger(cr model.Cron) func() {
-	crIgnoreMap := make(map[uint64]bool)
-	for j := 0; j < len(cr.Servers); j++ {
-		crIgnoreMap[cr.Servers[j]] = true
+	if err != nil {
+		panic(err)
 	}
-	return func() {
-		ServerLock.RLock()
-		defer ServerLock.RUnlock()
-		for _, s := range ServerList {
-			if cr.Cover == model.CronCoverAll && crIgnoreMap[s.ID] {
-				continue
-			}
-			if cr.Cover == model.CronCoverIgnoreAll && !crIgnoreMap[s.ID] {
-				continue
-			}
-			if s.TaskStream != nil {
-				s.TaskStream.Send(&pb.Task{
-					Id:   cr.ID,
-					Data: cr.Command,
-					Type: model.TaskTypeCommand,
-				})
-			} else {
-				SendNotification(fmt.Sprintf("[任务失败] %s，服务器 %s 离线，无法执行。", cr.Name, s.Name), false)
-			}
-		}
+	if Conf.Debug {
+		DB = DB.Debug()
 	}
-}
-
-func IPDesensitize(ip string) string {
-	if Conf.EnablePlainIPInNotification {
-		return ip
+	err = DB.AutoMigrate(model.Server{}, model.User{},
+		model.Notification{}, model.AlertRule{}, model.Monitor{},
+		model.MonitorHistory{}, model.Cron{}, model.Transfer{})
+	if err != nil {
+		panic(err)
 	}
-	return utils.IPDesensitize(ip)
 }
